Added two bug traps to write_report.pl
Added text to make it clear that in silico CMAP and key file can already exist if the in_silico directory if previous steps in the workflow added them and this is not an error. It is only a warning so that the user is aware in case the user has switched cmaps.

Added suggestion to fatal error when the alignment directory can't be found and the --de_novo flag was not used.